refactor(memory): reframe graph base as kernel substrate + adjacency-list impl per ADR (Refs #435)
Applies @titan's ADR ruling (adopted by the CEO): AbstractGraphMemory was
marketed as "generic CSR" but implemented a singly-linked adjacency list whose
neighbours() allocated a fresh int[] per call, violating the zero-alloc
read-path ethos.

- AbstractGraphMemory is now a thin kernel SUBSTRATE: SMKM 64-byte header,
  StampedLock SWMR, arena/segment/identity/shape/schema, WAL binding (inherited),
  close/flush skeleton, and the vertex-id space. It no longer implements the
  GraphMemory edge methods; concrete graphs own their layout.
- Extracted the bundled linked-list reference impl to AdjacencyListGraphMemory
  (VERTEX/EDGE slabs, free-lists, addNode/removeNode/addEdge/removeEdge/counts).
- neighbours() no longer allocates a per-call int[] snapshot: it returns a lazy
  iterator that walks the linked list under short validated read locks (the edge
  slab is never relocated, so slots stay stable).
- Moved edgeCapacity() out of the substrate onto AdjacencyListGraphMemory where
  it is truthful (the only caller was GraphMemoryContractTest, repointed).
- Reframed GraphMemory javadoc away from "CSR/slab" to adjacency graph wording.

HebbianGraphMemory (substrate mode) is unaffected; full module suite green
(1176 pass, 0 fail, 18 skipped).
